refactor: update log model and related functions for improved handling of log IDs
- Changed Log ID type from int to int64 for better compatibility with larger datasets.
- Updated log retrieval and creation functions to accommodate the new ID type.
- Refactored log trace handling to ensure consistent usage of int64 for log IDs.
- Enhanced database migration logic to support bigint columns for log IDs in PostgreSQL.
- Improved tests to validate the changes in log ID handling and ensure data integrity.

+++ model/usage_ledger.go
@@ -9,7 +9,7 @@ import (
 
 type UsageLedger struct {
 	Id                   int      `json:"id"`
-	NewapiLogId          int      `json:"newapi_log_id" gorm:"uniqueIndex"`
+	NewapiLogId          int64    `json:"newapi_log_id" gorm:"uniqueIndex"`
 	RequestId            string   `json:"request_id" gorm:"type:varchar(64);index;default:''"`
 	ExternalRequestId    string   `json:"external_request_id" gorm:"type:varchar(128);index;default:''"`
 	Timestamp            int64    `json:"timestamp" gorm:"bigint;index;index:idx_usage_ledger_vendor_profile_time,priority:2;index:idx_usage_ledger_channel_time,priority:2;index:idx_usage_ledger_model_time,priority:2;index:idx_usage_ledger_token_time,priority:2;index:idx_usage_ledger_provider_key_time,priority:2;index:idx_usage_ledger_group_time,priority:2;index:idx_usage_ledger_biz_scene_time,priority:3"`
